Allowed find to work and abstracted logic to seperate file
