+++ lib/customFiltering.js
@@ -1,0 +1,148 @@
+/**
+ * Gets extra filters for this resource object, which aren't part of the user-defined
+ * filters given with the filter query param. This function basically calls the (optional)
+ * filterMapper function defined in the resourceConfig, and passes the optional
+ * extraFilters object set in the request header.
+ *
+ * @param {Object} request - express.js request object
+ */
+const getExtraFilters = request => {
+  let extraFilters = {};
+  if (request.resourceConfig.filterMapper &&
+    typeof request.resourceConfig.filterMapper === 'function') {
+    extraFilters = request.resourceConfig.filterMapper(request.headers.extraFilters);
+    Object.entries(extraFilters).forEach(([key, value]) => {
+      extraFilters[key] = [ { operator: null, value } ]
+    });
+  }
+  return extraFilters;
+};
+
+/**
+ * Checks whether an attribute defines a relationship for a resource.
+ * Apparently, this can be checked via the _settings property.
+ * (got this from original _getSearchCriteria implementation)
+ *
+ * @param {String} attribute - Name of an attribute of a filter 
+ */
+const isRelationship = (resourceConfig, attribute) => {
+  const attributeConfig = resourceConfig.attributes[attribute];
+  return attributeConfig && attributeConfig._settings;
+};
+
+/**
+ * If given a valid string id, will cast it to a mongo ObjectId.
+ *
+ * @param {String} id - Id of a relationship
+ */
+const castRelationshipId = id => {
+  if (mongodb.ObjectId.isValid(id)) {
+    return new mongodb.ObjectId(id);
+  }
+  return id;
+};
+
+/**
+ * Recursive function which runs through a criteria object. 
+ * This criteria object can contain strings, arrays or objects, potentially 
+ * with mongo expressions sprinkled throughout. After running through the 
+ * recursion, the returned criteria will have all it's valid id's which map 
+ * to a relationship on the resource converted to a valid mongodb ObjectId.
+ *
+ * @param {String} key - The key of the criteria attribute
+ * @param {*} val - The value of the criteria attribute
+ * @param {Object} resourceConfig - Configuration details of the resource,
+ *                                  defined in it's index.js
+ */
+const parseRelationshipAttributes = (key, val, resourceConfig) => {
+  if (Array.isArray(val)) {
+    return val.map(subVal => parseRelationshipAttributes(key, subVal, resourceConfig));
+  }
+  if (typeof val === 'object') {
+    if (val['_bsontype']) {
+      // Allready an object id if it has a _bsontype
+      return val;
+    }
+
+    let parsedVal = val;
+    Object.keys(val).forEach(valKey => {
+      let keyToCheck = valKey;
+      if (valKey.substring(0, 1) === '$') {
+        // mongo operator, ignore it as a level for relationship key check in the recursion
+        keyToCheck = key;
+      }
+      parsedVal[valKey] = parseRelationshipAttributes(keyToCheck, val[valKey], resourceConfig);
+    });
+    return parsedVal;
+  }
+
+  // value is not an object or array. Actual relationship parsing possible.
+  if (isRelationship(resourceConfig, key)) {
+    return castRelationshipId(val);
+  }
+
+  return val;
+};
+
+/**
+ * Parses a criteria object so all relationship ids are valid mongodb ObjectIds. 
+ *
+ * @param {Object} resourceConfig - Configuration details of the resource,
+ *                                  defined in it's index.js
+ * @param {*} criteria - Generated criteria by the _getSearchCriteria function
+ */
+const parseCriteria = (resourceConfig, criteria) => {
+  console.log(JSON.stringify(criteria));
+  const parsedCriteria = {};
+  Object.keys(criteria).forEach(key => {
+    parsedCriteria[key] = parseRelationshipAttributes(key, criteria[key], resourceConfig);
+  });
+  return parsedCriteria;
+};
+
+/**
+ * Exposed the parsing to a mongo expression part of the _getSearchCriteria
+ * function so it can be easily reused in the find function.
+ *
+ * @param {Object} filter 
+ */
+parseCriteriaToMongoExpr = (MongoStore, filter) => {
+  const criteria = Object.keys(filter)
+    .map(attribute => {
+      let values = filter[attribute].map(MongoStore._filterElementToMongoExpr);
+      values = values.reduce((mongoExpressions, mongoExpr) => {
+        if (mongoExpr !== null) {
+          const mongoExprForAttr = {};
+          mongoExprForAttr[attribute] = mongoExpr;
+          mongoExpressions.push(mongoExprForAttr);
+        }
+        return mongoExpressions;
+      }, []);
+      if (values.length === 0) {
+        return null;
+      }
+      if (values.length === 1) {
+        return values[0];
+      }
+      return {
+        $or: values,
+      };
+    })
+    .filter(value => value !== null);
+
+  if (criteria.length === 0) {
+    return { };
+  }
+  if (criteria.length === 1) {
+    return criteria[0];
+  }
+  return {
+    $and: criteria,
+  };
+};
+
+module.exports = {
+  getExtraFilters,
+  parseCriteria,
+  parseCriteriaToMongoExpr,
+};
