+++ lib/mongoHandler.js
@@ -1,15 +1,15 @@
-/* eslint-disable */
-'use strict';
-var _ = require('lodash');
-var async = require('async');
-var mongodb = require('mongodb');
-var Joi = require('joi');
-var semver = require('semver');
-var debug = require('./debugging');
+const _ = require('lodash');
+const async = require('async');
+const mongodb = require('mongodb');
+const Joi = require('joi');
+const semver = require('semver');
+const debug = require('./debugging');
 
-var MIN_SERVER_VERSION = '1.10.0';
+const { getExtraFilters, parseCriteria, parseCriteriaToMongoExpr } = require('./customFiltering');
 
-var MongoStore = (module.exports = function MongoStore(config) {
+const MIN_SERVER_VERSION = '1.10.0';
+
+const MongoStore = (module.exports = function MongoStore(config) {
   MongoStore._checkMinServerVersion();
   this._config = config;
 });
@@ -80,50 +80,9 @@ MongoStore._filterElementToMongoExpr = function(filterElement) {
  * @private
  */
 MongoStore.prototype._getSearchCriteria = request => {
-  let extraFilters = {};
-  if (request.headers.extraFilters &&
-    request.resourceConfig.filterMapper &&
-    typeof request.resourceConfig.filterMapper === 'function') {
-    extraFilters = request.resourceConfig.filterMapper(request.headers.extraFilters);
-    Object.entries(extraFilters).forEach(([key, value]) => {
-      extraFilters[key] = [ { operator: null, value } ]
-    });
-  }
+  const filter = { ...(request.processedFilter || {}), ...getExtraFilters(request) };
 
-  const filter = Object.assign(request.processedFilter || {}, extraFilters);
-
-  const criteria = Object.keys(filter)
-    .map(attribute => {
-      let values = filter[attribute].map(MongoStore._filterElementToMongoExpr);
-      values = values.reduce((mongoExpressions, mongoExpr) => {
-        if (mongoExpr !== null) {
-          const mongoExprForAttr = {};
-          mongoExprForAttr[attribute] = mongoExpr;
-          mongoExpressions.push(mongoExprForAttr);
-        }
-        return mongoExpressions;
-      }, []);
-      if (values.length === 0) {
-        return null;
-      }
-      if (values.length === 1) {
-        return values[0];
-      }
-      return {
-        $or: values,
-      };
-    })
-    .filter(value => value !== null);
-
-  if (criteria.length === 0) {
-    return { };
-  }
-  if (criteria.length === 1) {
-    return criteria[0];
-  }
-  return {
-    $and: criteria,
-  };
+  return parseCriteriaToMongoExpr(MongoStore, filter);
 };
 
 MongoStore._notFoundError = function(type, id) {
@@ -266,60 +225,7 @@ MongoStore.prototype.populate = function(callback) {
  */
 MongoStore.prototype.search = function(request, callback) {
   const collection = this._db.collection(request.params.type);
-  let criteria = this._getSearchCriteria(request);
-
-  console.log(criteria);
-  // Checks whether an attribute defines a relationship for a resource.
-  // Apparently, this can be checked via the _settings property.
-  // (got this from original _getSearchCriteria implementation)
-  const isRelationship = attribute => {
-    const attributeConfig = this.resourceConfig.attributes[attribute];
-    return attributeConfig && attributeConfig._settings;
-  };
-
-  // If given a valid string id, will cast it to a mongo ObjectId.
-  const castRelationshipId = id => {
-    if (mongodb.ObjectId.isValid(id)) {
-      return new mongodb.ObjectId(id);
-    }
-    return id;
-  };
-
-  const parseRelationshipAttributes = (key, val) => {
-    if (Array.isArray(val)) {
-      return val.map(subVal => parseRelationshipAttributes(key, subVal));
-    }
-    if (typeof val === 'object') {
-      if (val['_bsontype']) {
-        // Allready an object id if it has a _bsontype
-        return val;
-      }
-
-      let parsedVal = val;
-      Object.keys(val).forEach(valKey => {
-        let keyToCheck = valKey;
-        if (valKey.substring(0, 1) === '$') {
-          // mongo operator, ignore it as a level for relationship key check in the recursion
-          keyToCheck = key;
-        }
-        parsedVal[valKey] = parseRelationshipAttributes(keyToCheck, val[valKey]);
-      });
-      return parsedVal;
-    }
-
-    // value is not an object or array. Actual relationship parsing possible.
-    if (isRelationship(key)) {
-      return castRelationshipId(val);
-    }
-
-    return val;
-  };
-
-  Object.keys(criteria).forEach(key => {
-    criteria[key] = parseRelationshipAttributes(key, criteria[key]);
-  });
-
-  console.log(JSON.stringify(criteria));
+  let criteria = parseCriteria(this.resourceConfig, this._getSearchCriteria(request));
 
   async.parallel(
     {
@@ -344,19 +250,22 @@ MongoStore.prototype.search = function(request, callback) {
   Find a specific resource, given a resource type and and id.
  */
 MongoStore.prototype.find = function(request, callback) {
-  var collection = this._db.collection(request.params.type);
-  var documentId = MongoStore._mongoUuid(request.params.id);
+  const collection = this._db.collection(request.params.type);
+  const documentId = MongoStore._mongoUuid(request.params.id);
+  const filter = parseCriteriaToMongoExpr(MongoStore, getExtraFilters(request));
+  const criteria = parseCriteria(this.resourceConfig,filter);
 
   debug(
     'findOne',
     {
       _id: documentId,
+      ... criteria,
     }
   );
   collection.findOne(
     {
       _id: documentId,
-      isSoftDeleted: false,
+      ... criteria,
     },
     function(err, result) {
       if (err || !result) {
